feat: implement DashboardTemplateEditor component for managing dashboard templates
Added a new DashboardTemplateEditor component to facilitate the creation and editing of dashboard templates. This component integrates with the existing dashboard template service and includes form validation and state management. Updated the DashboardTemplateManager to utilize this new component, streamlining the template management process. Additionally, modified the save functionality to ensure proper handling of widget configurations.
